pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.0.1"
description = "\"A customized djagno template to use any projet\""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poe.tasks]
dev = "python manage.py runserver 0.0.0.0:8000"
migrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
app = "python manage.py startapp"

[tool.poetry.dependencies]

# Python dependencies
python = "^3.12"

# Django dependencies
django = "^5.1.2"
pillow = "^11.0.0"
django-cors-headers = "^4.6.0"
django-phonenumber-field = "^8.0.0"
django-admin-rangefilter = "^0.13.2"

# pydantic dependencies
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
pydantic-extra-types = "^2.9.0"

#logging dependencies
loguru = "^0.7.2"

# development dependencies
ruff = "^0.7.1"
django-debug-toolbar = "^4.4.6"
pre-commit = "^4.0.1"

# rest dependencies
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"

# api documentation dependencies
drf-spectacular = "^0.27.2"

# server dependencies
granian = "^1.6.3"
uvicorn = "^0.32.0"

# cashe dependencies
redis = "^5.2.0"

# Others
psycopg = "^3.2.3"
pyseto = "^1.7.9"
whitenoise = "^6.8.2"
phonenumbers = "^8.13.48"
poethepoet = "^0.29.0"



[tool.poetry.group.dev.dependencies]
poethepoet = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
